Clear CFLAGS CXXFLAGS LDFLAGS.
This is a bit of a hammer, but without this the regtests results are just
horrible with hundreds of failures. Even with this there are 29 failures.
Once we fix those and have a clean testsuite we should reinstate the
FLAGS and figure out exactly which ones cause the massive fails.