Enable regtest suite on ARM.
Make sure make check gets empty CFLAGS so the architecture specific ones
don't clash. And remove the annotate_trace_memory tests on ARM.
There is a __sync_add_and_fetch in the testcase. DRD is doing this trace
printing inside the loop which causes the reservation (LDREX) to fail so
it can never make progress.