does-Image-Magick-works-under-valgrind test to Fedora CI
The original RHEL test also converts an image to png. This version
of the test has the conversion to png disabled as it faills the test.
The failure is probably caused by some rounding issue which was
impossible to track down so far.

The test is currently only run on x86_64. There might be "Conditional jump
or move depends on uninitialised value(s)" warnings while running under valgrind,
but the test is to show the transformations work with and without valgrind.